Added make_pair function and moved transliteration file
