+++ src/bot/models.py
@@ -2,7 +2,7 @@ from django.db import models
 from django.db.models.signals import pre_save
 from django.dispatch import receiver
 
-from .utils import transliteration
+from bot.utils.transliteration import transliteration
 
 
 class Profession(models.Model):
